04-ZABURO/code/submodules/somen/requirements.txt
pysen[lint]
pandas
numpy
tables
torch
scikit-learn
pfio
albumentations
pytorch-pfn-extras
torchvision
feather-format
dacite
colorlog
tqdm
